% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{lower_pi}
\alias{lower_pi}
\title{Return lower boundary for the percentile interval}
\usage{
lower_pi(values, CI = 0.97)
}
\arguments{
\item{values}{Vector of numeric values}

\item{CI}{Float, confidence level. Default is `0.97`}
}
\value{
Float
}
\description{
Return lower boundary for the percentile interval
}
