% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_normal}
\alias{standard_normal}
\title{Standard normally distributed values
Table with normally distributed values
with mean = 0 and standard deviation = 1.}
\format{

}
\usage{
standard_normal
}
\description{
Standard normally distributed values
Table with normally distributed values
with mean = 0 and standard deviation = 1.
}
\keyword{datasets}
